Implement gateway logic
Added the handling of serial console messages for "GET" and
"POST" requests, by parsing the received message into a JSON.

The messages written on console by the micro:bit are read and
then a request is generated based on the METHOD found inside,
which is forwarded to the server. The request response is
written back by the gateway app to serial console, which
the micro:bit should read.

Signed-off-by: Andrei Cherechesu <andrei.cherechesu@gmail.com>